Implement SQL Query and method to use it
Create SQL Query to count the total of vehicle reg number are in the
databases (DBConstants).

Create method countVehicleRegNumber in class ParkingSpotDAO to execute
the Query and return the result

Implement the method regularCustomer in class ParkingService to return
true if the vehicle reg number already exist in the database.

This method is call by
ticket.setDiscount(regularCustomer(vehicleRegNumber)); in the method
processIncomingVehicule in the class ParkinService line 44.